requirements.txt
torch>=2.0.0
hydra-core>=1.3.0
wandb>=0.15.0
numpy>=1.24.0
einops>=0.7.0
tqdm>=4.65.0

